refactor: 4단계 - 불필요 코드 정리
