+++ README.md
@@ -99,7 +99,7 @@ TDD로 구현
   - "|     |------|     |"
   - "|-----|      |-----|"
 
-- [ ] 결과 조회
+- [x] 결과 조회
 
 
 Usernames
