+++ src/main/java/nextstep/ladder/Ladder.java
@@ -1,64 +1,53 @@
 package nextstep.ladder;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class Ladder {
-    private Usernames usernames;
-    private Lines lines;
-    private List<String> results;
+    private static final String LADDER_DELIMITER = "\n";
 
-    public Ladder(Usernames usernames, Lines lines, List<String> results) {
-        this.usernames = usernames;
-        this.lines = lines;
-        this.results = results;
+    private List<Line> lines;
+
+    public Ladder(int countOfPerson, int height) {
+        this.lines = createLines(countOfPerson, height);
     }
 
-    public String getResultsForPrint() {
-        return results.stream()
-                .map(d -> String.format("%-6s", d))
-                .collect(Collectors.joining());
+    public Ladder(List<Line> lines) {
+        this.lines = new ArrayList<>(lines);
+    }
+
+    private List<Line> createLines(int countOfPerson, int height) {
+        List<Line> nLines = new ArrayList<>();
+
+        for (int i = 0; i < height; i++) {
+            nLines.add(new Line(countOfPerson));
+        }
+
+        return nLines;
+    }
+
+    public int move(int position) {
+        int index = position;
+        for (Line line : lines) {
+            index = line.move(index);
+        }
+        return index;
     }
 
     public String getLadderForPrint(String trueSymbol, String falseSymbol) {
-        return String.join("\n",
-                usernames.toString(),
-                lines.toString(trueSymbol, falseSymbol),
-                getResultsForPrint()
-        );
+        return lines.stream()
+                .map(line -> line.getLineForPrint(trueSymbol, falseSymbol))
+                .collect(Collectors.joining(LADDER_DELIMITER));
     }
 
-    public String getResult(String resultUsername) {
-        if ("all".equals(resultUsername)) {
-            return getAllResult();
-        }
-
-        return getResultByUsername(resultUsername);
-    }
-
-    private Map<String, String> getResultMap() {
-        Map<String, String> result = IntStream.range(0, usernames.size())
+    //todo lines.get(0).getCrossSize()
+    public Map<Integer, Integer> getAllResult() {
+        return IntStream.range(0, lines.get(0).getCrossSize())
                 .boxed()
-                .collect(Collectors.toMap(
-                        i -> usernames.getValues().get(i).getUsername(),
-                        i -> results.get(lines.getPersonFinalResultIndex(i))
-                ));
-
-        return result;
-    }
-
-    private String getAllResult() {
-        Map<String, String> resultMap = getResultMap();
-
-        return resultMap.entrySet()
-                .stream()
-                .map(d -> d.getKey() + " : " + d.getValue()).collect(Collectors.joining("\n"));
-    }
-
-    private String getResultByUsername(String resultUsername) {
-        return getResultMap().get(resultUsername);
+                .collect(Collectors.toMap(i -> i, i -> move(i)));
     }
 
 }
