+++ src/main/java/nextstep/ladder/LadderMain.java
@@ -9,10 +9,12 @@ public class LadderMain {
         if (usernames.size() != results.size()) {
             throw new IllegalArgumentException("참여할 사람 수와 입력한 실행 결과의 수가 다릅니다.");
         }
-        Ladder ladder = new Ladder(usernames, new Lines(InputView.inputHeight(), usernames.size()), results);
+        Ladder ladder = new Ladder(usernames.size(), InputView.inputHeight());
+        ResultView.printUsernames(usernames);
         ResultView.printLadder(ladder);
+        ResultView.printResults(results);
 
         String resultUserName = InputView.inputResultUserName();
-        ResultView.printResult(ladder.getResult(resultUserName));
+        ResultView.printResult(resultUserName, usernames, ladder, results);
     }
 }
