--- src/main/java/nextstep/ladder/Line2.java
@@ -1,53 +1,0 @@
-package nextstep.ladder;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
-import java.util.stream.Collectors;
-
-public class Line2 {
-    private static final Random RANDOM = new Random();
-    private static final String LINE_DELIMITER = "|";
-
-    private List<Cross> crosses;
-
-    public Line2(int countOfPerson) {
-        createCrosses(countOfPerson);
-    }
-
-    public Line2(List<Cross> crosses) {
-        this.crosses = crosses;
-    }
-
-    private void createCrosses(int countOfPerson) {
-        crosses = new ArrayList<Cross>();
-
-        Point point = Point.first(RANDOM.nextBoolean());
-        for (int i = 0; i < countOfPerson; i++) {
-            if (i == countOfPerson - 1) {
-                point = point.last();
-            }
-            if (i > 0 && i < countOfPerson - 1) {
-                point = point.next(RANDOM.nextBoolean());
-            }
-            crosses.add(new Cross(i, point));
-        }
-    }
-
-    //todo
-    public int move(int position) {
-        return this.crosses.get(position).move();
-    }
-
-    public String getLineForPrint(String trueSymbol, String falseSymbol) {
-        return LINE_DELIMITER
-                + crosses.stream()
-                .map(cross -> cross.getCrossForPrint(trueSymbol, falseSymbol))
-                .collect(Collectors.joining(LINE_DELIMITER));
-    }
-
-    //todo
-    public int getCrossSize() {
-        return crosses.size();
-    }
-}
